Fix a circular dep cycle
