+++ pages/api/sites/[siteName].ts
@@ -1,14 +1,14 @@
 import { NextApiRequest, NextApiResponse } from "next";
 
 import env from "../../../lib/server/environment";
-import { InstantBanditServer } from "../../../server";
+import { InstantBanditServer } from "../../../lib/server/server-types";
 import { DEFAULT_SITE } from "../../../lib/defaults";
 import { exists } from "../../../lib/utils";
-import { validateUserRequest } from "../../../server";
+import { validateUserRequest } from "../../../lib/server/server-utils";
 import { getInternalDevServer } from "../../../lib/server/server-internal";
 
 
-const SiteEndpoint = createSiteEndpoint(getInternalDevServer());
+const SiteEndpoint = createSiteEndpoint();
 export default SiteEndpoint;
 
 /**
@@ -16,7 +16,7 @@ export default SiteEndpoint;
  * Remember to return the resulting endpoint function as your module's default export.
  * See [siteName].ts.
  */
-export function createSiteEndpoint(server: InstantBanditServer) {
+export function createSiteEndpoint(server?: InstantBanditServer) {
 
   // This endpoint serves site configurations by name and creates user sessions.
   // Site are hydrated with variant probabilities inlined.
@@ -25,6 +25,9 @@ export function createSiteEndpoint(server: InstantBanditServer) {
   // Session IDs are transmitted via a 1st-party cookie.
 
   async function handleSiteRequest(req: NextApiRequest, res: NextApiResponse) {
+    if (!server) {
+      server = getInternalDevServer();
+    }
     await server.init();
 
     // TODO: Respond to CORS preflights
